refactor(Calendar) moved setting dates to custom hook
